Add optimal goals assignment algorithm

+++ mapf_project/agent_meeting_planner.py
@@ -6,7 +6,7 @@ import sys
 from libraries.cbs import CBSSolver
 from libraries.connectivity_graphs import are_nodes_a_clique, generate_connectivity_graph, import_connectivity_graph, print_connectivity_graph
 from libraries.enums import ConnectionCriterion, GoalsChoice, GoalsAssignment
-from libraries.goals_assignment import print_goals_assignment, search_goals_assignment_astar, search_goals_assignment_exhaustive_search_astar
+from libraries.goals_assignment import print_goals_assignment, search_goals_assignment_astar, search_goals_assignment_exhaustive_search_astar, search_goals_assignment_cbs
 from libraries.goals_choice import get_distance_to_all_starting_locations, print_goal_positions
 from libraries.run_experiments import import_mapf_instance
 from libraries.single_agent_planner import compute_heuristics
@@ -81,6 +81,9 @@ def get_goals_assignment(map: list[list[bool]], starts: list[tuple[int, int]], g
     elif args.goals_assignment == GoalsAssignment.MINIMIZE_DISTANCE_ASTAR.name:
         new_goals = search_goals_assignment_astar(map, starts, goal_positions)
 
+    elif args.goals_assignment == GoalsAssignment.MINIMIZE_DISTANCE_CBS.name:
+        new_goals = search_goals_assignment_cbs(map, starts, goal_positions)
+
     else:
         raise(RuntimeError("Unknown goals assignment algorithm."))
 
@@ -88,11 +91,14 @@ def get_goals_assignment(map: list[list[bool]], starts: list[tuple[int, int]], g
 
 def solve_instance(file: str, args: list) -> None:
     if (args.save_output):
+        print("Solving " + file)
         file_name_sections = file.split("\\")
         file_id = file_name_sections[-1].split(".")[0]
         path = "./outputs/" + file_id + ".txt"
         f = open(path, 'w')
         sys.stdout = f
+
+    start_time = time.time()
 
     print("*** Import an instance ***\n")
     print("Instance: " + file + "\n")
@@ -121,6 +127,9 @@ def solve_instance(file: str, args: list) -> None:
     print("*** Modified problem ***\n")
     print_mapf_instance(map, starts, new_goals)
 
+    CPU_time = time.time() - start_time
+    print("CPU time (s):    {:.2f}".format(CPU_time))
+
     if (args.solve):
         print("***Run CBS***")
         cbs = CBSSolver(map, starts, new_goals)
@@ -135,7 +144,7 @@ if __name__ == '__main__':
                         help='The name of the instance file(s)')
     parser.add_argument('--goals_choice', type=str, default=GoalsChoice.GREEDY_MINIMIZE_DISTANCE.name, choices=[GoalsChoice.GREEDY_MINIMIZE_DISTANCE.name],
                         help='The algorithm to use to select the goal nodes, defaults to ' + GoalsChoice.GREEDY_MINIMIZE_DISTANCE.name)
-    parser.add_argument('--goals_assignment', type=str, default=GoalsAssignment.MINIMIZE_DISTANCE_ASTAR.name, choices=[GoalsAssignment.ARBITRARY.name, GoalsAssignment.RANDOM.name, GoalsAssignment.GREEDY.name, GoalsAssignment.EXHAUSTIVE_SEARCH_ASTAR.name, GoalsAssignment.MINIMIZE_DISTANCE_ASTAR.name],
+    parser.add_argument('--goals_assignment', type=str, default=GoalsAssignment.MINIMIZE_DISTANCE_ASTAR.name, choices=[GoalsAssignment.ARBITRARY.name, GoalsAssignment.RANDOM.name, GoalsAssignment.GREEDY.name, GoalsAssignment.EXHAUSTIVE_SEARCH_ASTAR.name, GoalsAssignment.MINIMIZE_DISTANCE_ASTAR.name, GoalsAssignment.MINIMIZE_DISTANCE_CBS.name],
                         help='The algorithm to use to assign each goal to an agent, defaults to ' + GoalsAssignment.MINIMIZE_DISTANCE_ASTAR.name)
     parser.add_argument('--connectivity_graph', type=str, default=None,
                         help='The name of the file containing the connectivity graph, if included it will be imported from said file instead of being generated')
@@ -147,6 +156,8 @@ if __name__ == '__main__':
                         help='Decide to solve the instance using CBS or not, defaults to ' + str(False))
     parser.add_argument('--save_output', type=bool, default=False,
                         help='Decide to save the output in txt files, defaults to ' + str(False))
+    parser.add_argument('--timeout', type=bool, default=False,
+                        help='Decide to terminate the solver if it cannot finish in less than a minute, defaults to ' + str(False))
 
     args = parser.parse_args()
 
@@ -154,13 +165,15 @@ if __name__ == '__main__':
         p = multiprocessing.Process(target=solve_instance, name="Solve instance", args=(file, args))
         p.start()
 
-        counter = 0
-        while (counter < SOLVER_TIMEOUT):
-            time.sleep(1)
-            counter += 1
-            if (not p.is_alive()):
-                break
-        if (p.is_alive()):
-            print("Solver coult not finish in " + str(SOLVER_TIMEOUT) + " seconds and was terminated")
-            p.terminate()
-            p.join()
+        if (args.timeout):
+            counter = 0
+            while (counter < SOLVER_TIMEOUT):
+                time.sleep(1)
+                counter += 1
+                if (not p.is_alive()):
+                    break
+            if (p.is_alive()):
+                print("Solver coult not finish in " + str(SOLVER_TIMEOUT) + " seconds and was terminated")
+                p.terminate()
+
+        p.join()
